- Updated .gitignore - Updated Liquid.Json.Mvc project to not depend tightly on the version of MVC - Added StyleCop Settings - Ran ReSharper Code Cleanup across solution

+++ tests/Liquid.Json.Tests/Writer/Integers.cs
@@ -1,25 +1,27 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.IO;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Liquid.Json.Tests.Writer {
+namespace Liquid.Json.Tests.Writer
+{
     [TestClass]
-    public class Integers {
+    public class Integers
+    {
         [TestMethod, TestCategory("Writer")]
-        public void WritesInt32() {
+        public void WritesInt32()
+        {
             var sw = new StringWriter();
             var writer = new JsonWriter(sw);
-            writer.WriteValue((Int32)123);
+            writer.WriteValue(123);
             Assert.AreEqual("123", sw.ToString());
         }
+
         [TestMethod, TestCategory("Writer")]
-        public void WritesInt16() {
+        public void WritesInt16()
+        {
             var sw = new StringWriter();
             var writer = new JsonWriter(sw);
-            writer.WriteValue((Int16)123);
+            writer.WriteValue((Int16) 123);
             Assert.AreEqual("123", sw.ToString());
         }
     }
